move getSourceCode outside
