+++ packages/config/scripts/checkVerifiedContracts/types.ts
@@ -1,3 +1,22 @@
+import { z } from 'zod'
 import { Bridge, Layer2, Layer3 } from '../../src'
 
 export type Project = Layer2 | Layer3 | Bridge
+
+export type ContractSource = z.infer<typeof ContractSource>
+export const ContractSource = z.object({
+  SourceCode: z.string(),
+  ABI: z.string(),
+  ContractName: z.string(),
+  CompilerVersion: z.string(),
+  OptimizationUsed: z.string(),
+  Runs: z.string(),
+  ConstructorArguments: z.string(),
+  EVMVersion: z.string(),
+  Library: z.string(),
+  LicenseType: z.string(),
+  Proxy: z.string(),
+  Implementation: z.string(),
+  SwarmSource: z.string(),
+})
+export const ContractSourceResult = z.array(ContractSource).length(1)
